Remove InfiniteScroll from the result page
This used to make scrolling impossible. The End key did not go to the
end of the list but rather to the middle. In case of a relay active say
from 2010 to 2019, End key did not go to the present as a sane person
would expect but rather somewhere to the middle of 2010. My expection is
that moving to the 2019 could take as much as an hour or maybe more.